# db: add LazyValue for a value that may not be stored in-place with th…