--- a/error_iter.go
+++ b/error_iter.go
@@ -20,34 +20,34 @@ func newErrorIter(err error) *errorIter {
 	return &errorIter{err: err}
 }
 
-func (c *errorIter) SeekGE(key []byte, flags base.SeekGEFlags) (*InternalKey, []byte) {
-	return nil, nil
+func (c *errorIter) SeekGE(key []byte, flags base.SeekGEFlags) (*InternalKey, base.LazyValue) {
+	return nil, base.LazyValue{}
 }
 
 func (c *errorIter) SeekPrefixGE(
 	prefix, key []byte, flags base.SeekGEFlags,
-) (*base.InternalKey, []byte) {
-	return nil, nil
+) (*base.InternalKey, base.LazyValue) {
+	return nil, base.LazyValue{}
 }
 
-func (c *errorIter) SeekLT(key []byte, flags base.SeekLTFlags) (*InternalKey, []byte) {
-	return nil, nil
+func (c *errorIter) SeekLT(key []byte, flags base.SeekLTFlags) (*InternalKey, base.LazyValue) {
+	return nil, base.LazyValue{}
 }
 
-func (c *errorIter) First() (*InternalKey, []byte) {
-	return nil, nil
+func (c *errorIter) First() (*InternalKey, base.LazyValue) {
+	return nil, base.LazyValue{}
 }
 
-func (c *errorIter) Last() (*InternalKey, []byte) {
-	return nil, nil
+func (c *errorIter) Last() (*InternalKey, base.LazyValue) {
+	return nil, base.LazyValue{}
 }
 
-func (c *errorIter) Next() (*InternalKey, []byte) {
-	return nil, nil
+func (c *errorIter) Next() (*InternalKey, base.LazyValue) {
+	return nil, base.LazyValue{}
 }
 
-func (c *errorIter) Prev() (*InternalKey, []byte) {
-	return nil, nil
+func (c *errorIter) Prev() (*InternalKey, base.LazyValue) {
+	return nil, base.LazyValue{}
 }
 
 func (c *errorIter) Error() error {