--- a/internal/metamorphic/ops.go
+++ b/internal/metamorphic/ops.go
@@ -467,7 +467,7 @@ func (o *ingestOp) build(t *test, h historyRecorder, b *pebble.Batch, i int) (st
 		lastUserKey = key.UserKey
 
 		key.SetSeqNum(0)
-		if err := w.Add(*key, value); err != nil {
+		if err := w.Add(*key, value.InPlaceValue()); err != nil {
 			return "", err
 		}
 	}
@@ -565,9 +565,9 @@ func (o *ingestOp) collapseBatch(
 			case pebble.InternalKeyKindSingleDelete:
 				err = collapsed.SingleDelete(key.UserKey, nil)
 			case pebble.InternalKeyKindSet:
-				err = collapsed.Set(key.UserKey, value, nil)
+				err = collapsed.Set(key.UserKey, value.InPlaceValue(), nil)
 			case pebble.InternalKeyKindMerge:
-				err = collapsed.Merge(key.UserKey, value, nil)
+				err = collapsed.Merge(key.UserKey, value.InPlaceValue(), nil)
 			case pebble.InternalKeyKindLogData:
 				err = collapsed.LogData(key.UserKey, nil)
 			default: