--- a/mem_table.go
+++ b/mem_table.go
@@ -336,7 +336,7 @@ func (f *keySpanFrags) get(
 		it := skl.NewIter(nil, nil)
 		var keysDst []keyspan.Key
 		for key, val := it.First(); key != nil; key, val = it.Next() {
-			s, err := constructSpan(*key, val, keysDst)
+			s, err := constructSpan(*key, val.InPlaceValue(), keysDst)
 			if err != nil {
 				panic(err)
 			}