…e key

The changes here are plumbing for the read path. The actual write and read path for values stored in value blocks will happen in future PRs.

A ShortAttribute (3 bits) is also introduced, for user-defined attributes about a value. These will be included in the 1 byte prefix of the value in the key-value pairs in the sstable (the other 5 bits are for internal use). In CockroachDB, the ShortAttribute will initially be used for marking MVCC tombstones (for which 1 bit is sufficient).

LazyValue can represent either an in-place value (stored with the key) or a value stored in the value section. The interface is general enough for the future where values may be stored in separate files.

LazyValue is used in the InternalIterator interface, such that all positioning calls return (*InternalKey, LazyValue). It is also exposed via the public Iterator for callers that need to remember a recent but not necessarily latest LazyValue, in case they need the actual value in the future. An example is a caller that is iterating in reverse and looking for the latest MVCC version for a key -- it cannot identify the latest MVCC version without stepping to the previous key-value pair e.g. storage.pebbleMVCCScanner in CockroachDB.

There is subtlety in memory management:
- A LazyValue returned by an InternalIterator or Iterator is unstable in that repositioning the iterator will invalidate the memory inside it. A caller wishing to maintain that LazyValue needs to call LazyValue.Clone. Note that this does not fetch the value if it is not in-place.

- A user of an iterator that calls LazyValue.Value wants as much as possible for the returned value byte-slice to point to iterator owned memory.
  - [P1] The underlying iterator that owns that memory also needs a promise from that user that at any time there is at most one value byte-slice that the caller is expecting it to maintain. Otherwise, the underlying iterator has to maintain multiple such byte-slices which results in more complicated and inefficient code. Note that pre-LazyValue this promise was trivially provided by only asking for the value at the current iterator position (so it could certainly point inside the underlying ssblock). For other cases, the user would make a copy of the byte-slice. We are generalizing this promise which allows for more memory allocation optimization.
  - [P2] The underlying iterator, in order to make the promise that it is maintaining the one value byte-slice, also needs a way to know when it is relieved of that promise. It is relieved of that promise by being told that it is being repositioned (this behavior is unchanged by this PR). Specifically, the owner of the value byte-slice is a sstable iterator. Consider the case where the caller has used LazyValue.Clone and repositioned the iterator. In this case the sstable iterator may not even be open. LazyValue.Value will still work (at a higher cost), but since the sstable iterator is not open, it cannot satisfy P2. For this reason, the LazyValue.Value method accepts a caller owned buffer, that the callee will use if needed.