@sumeerbhola 2022-10-14
The Pebble benchmark numbers are slightly slower, but they are a bit noisy in that another run on my gceworker had master slower than this PR. The only consistently slower one is `IteratorScan/keys=1000,r-amp=1,key-types=points-and-ranges-24                 85.1µs ± 1%   107.6µs ±24%  +26.45%  (p=0.008 n=5+5)` which I will investigate.

```
name                                                                        old time/op    new time/op    delta
pkg:github.com/cockroachdb/pebble goos:linux goarch:amd64
IteratorSeekGE-24                                                              986ns ± 2%    1011ns ± 8%     ~     (p=0.690 n=5+5)
IteratorNext-24                                                               33.1ns ± 2%    34.5ns ± 1%   +4.15%  (p=0.008 n=5+5)
IteratorPrev-24                                                               45.8ns ± 1%    43.6ns ± 1%   -4.77%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/bloom=false/with-tombstone=false-24     694ns ± 2%     712ns ± 1%   +2.70%  (p=0.016 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/bloom=false/with-tombstone=true-24      452ns ± 1%     458ns ± 0%   +1.35%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/bloom=true/with-tombstone=false-24     1.32µs ± 0%    1.35µs ± 0%   +1.62%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/bloom=true/with-tombstone=true-24      1.98µs ± 1%    1.99µs ± 1%     ~     (p=0.548 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/bloom=false/with-tombstone=false-24     703ns ± 1%     720ns ± 1%   +2.42%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/bloom=false/with-tombstone=true-24      454ns ± 1%     459ns ± 2%     ~     (p=0.056 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/bloom=true/with-tombstone=false-24     1.33µs ± 1%    1.36µs ± 1%   +2.52%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/bloom=true/with-tombstone=true-24      1.98µs ± 1%    2.02µs ± 1%   +2.10%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/bloom=false/with-tombstone=false-24     709ns ± 1%     725ns ± 1%   +2.32%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/bloom=false/with-tombstone=true-24      453ns ± 3%     460ns ± 3%     ~     (p=0.151 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/bloom=true/with-tombstone=false-24     1.34µs ± 1%    1.36µs ± 1%   +1.31%  (p=0.040 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/bloom=true/with-tombstone=true-24      2.01µs ± 1%    2.02µs ± 1%     ~     (p=0.206 n=5+5)
IteratorSeqSeekGEWithBounds-24                                                 741ns ± 1%     753ns ± 0%   +1.64%  (p=0.008 n=5+5)
IteratorSeekGENoop/withLimit=false-24                                         55.0ns ± 3%    57.9ns ± 1%   +5.24%  (p=0.008 n=5+5)
IteratorSeekGENoop/withLimit=true-24                                          96.6ns ± 0%    98.0ns ± 0%   +1.41%  (p=0.016 n=4+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward-24                   5.03ms ± 2%    5.14ms ± 5%     ~     (p=0.222 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/backward-24                  6.35ms ± 4%    6.49ms ± 2%     ~     (p=0.222 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward-24                   4.56ms ± 2%    4.66ms ± 2%     ~     (p=0.056 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/backward-24                  5.58ms ± 1%    5.77ms ± 3%   +3.49%  (p=0.016 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward-24                   2.47ms ± 2%    2.52ms ± 2%     ~     (p=0.056 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/backward-24                  3.15ms ± 3%    3.24ms ± 4%     ~     (p=0.151 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward-24                   342µs ± 2%     351µs ± 2%   +2.80%  (p=0.032 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/backward-24                  669µs ± 2%     689µs ± 2%     ~     (p=0.056 n=5+5)
IteratorScan/keys=100,r-amp=1,key-types=points-only-24                        10.6µs ± 3%    10.7µs ± 2%     ~     (p=0.310 n=5+5)
IteratorScan/keys=100,r-amp=1,key-types=points-and-ranges-24                  11.3µs ±13%    11.4µs ± 3%     ~     (p=0.151 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-only-24                        19.7µs ±12%    20.0µs ± 5%     ~     (p=0.421 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-and-ranges-24                  20.0µs ±14%    20.9µs ±10%     ~     (p=0.095 n=5+5)
IteratorScan/keys=100,r-amp=7,key-types=points-only-24                        27.9µs ± 8%    27.6µs ± 1%     ~     (p=0.690 n=5+5)
IteratorScan/keys=100,r-amp=7,key-types=points-and-ranges-24                  29.3µs ±10%    29.2µs ± 8%     ~     (p=1.000 n=5+5)
IteratorScan/keys=100,r-amp=10,key-types=points-only-24                       35.4µs ± 3%    33.8µs ± 6%     ~     (p=0.095 n=5+5)
IteratorScan/keys=100,r-amp=10,key-types=points-and-ranges-24                 33.3µs ± 7%    35.4µs ± 5%     ~     (p=0.056 n=5+5)
IteratorScan/keys=1000,r-amp=1,key-types=points-only-24                       82.4µs ± 1%    85.5µs ± 2%   +3.74%  (p=0.016 n=5+4)
IteratorScan/keys=1000,r-amp=1,key-types=points-and-ranges-24                 85.1µs ± 1%   107.6µs ±24%  +26.45%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-only-24                        142µs ± 2%     151µs ± 3%   +6.25%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-and-ranges-24                  146µs ± 2%     154µs ± 1%   +5.83%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=7,key-types=points-only-24                        187µs ± 1%     196µs ± 2%   +4.46%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=7,key-types=points-and-ranges-24                  193µs ± 5%     205µs ± 7%     ~     (p=0.056 n=5+5)
IteratorScan/keys=1000,r-amp=10,key-types=points-only-24                       206µs ± 3%     214µs ± 6%     ~     (p=0.095 n=5+5)
IteratorScan/keys=1000,r-amp=10,key-types=points-and-ranges-24                 207µs ± 2%     221µs ± 2%   +6.68%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=1,key-types=points-only-24                       805µs ± 2%     838µs ± 1%   +4.16%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=1,key-types=points-and-ranges-24                 844µs ± 1%     902µs ± 2%   +6.77%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-only-24                      1.38ms ± 2%    1.44ms ± 3%   +4.56%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-and-ranges-24                1.41ms ± 5%    1.50ms ± 1%   +6.27%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=7,key-types=points-only-24                      1.75ms ± 3%    1.81ms ± 2%   +3.37%  (p=0.032 n=5+5)
IteratorScan/keys=10000,r-amp=7,key-types=points-and-ranges-24                1.79ms ± 4%    1.89ms ± 0%   +5.69%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-only-24                     1.90ms ± 1%    1.98ms ± 3%   +4.56%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-and-ranges-24               1.94ms ± 1%    2.04ms ± 1%   +4.74%  (p=0.008 n=5+5)
CombinedIteratorSeek/range-key=false/batch=false-24                           5.48µs ± 2%    5.74µs ± 1%   +4.80%  (p=0.008 n=5+5)
CombinedIteratorSeek/range-key=false/batch=true-24                            6.14µs ± 3%    6.46µs ± 1%   +5.16%  (p=0.008 n=5+5)
CombinedIteratorSeek/range-key=true/batch=false-24                            13.3µs ± 2%    13.6µs ± 1%   +2.17%  (p=0.032 n=5+5)
CombinedIteratorSeek/range-key=true/batch=true-24                             14.1µs ± 1%    14.5µs ± 1%   +2.59%  (p=0.008 n=5+5)
CombinedIteratorSeek_Bounded-24                                               94.4µs ± 1%    94.3µs ± 3%     ~     (p=1.000 n=5+5)
CombinedIteratorSeekPrefix-24                                                  184µs ± 4%     188µs ± 2%     ~     (p=0.056 n=5+5)
LevelIterSeekGE/restart=16/count=5-24                                         1.74µs ± 4%    1.70µs ± 1%     ~     (p=0.056 n=5+5)
LevelIterSeqSeekGEWithBounds/restart=16/count=5-24                             118ns ± 0%     121ns ± 0%   +2.64%  (p=0.008 n=5+5)
LevelIterSeqSeekPrefixGE/skip=1/use-next=false-24                              667ns ± 5%     658ns ± 1%     ~     (p=0.548 n=5+5)
LevelIterSeqSeekPrefixGE/skip=1/use-next=true-24                               131ns ± 0%     130ns ± 1%   -0.83%  (p=0.032 n=5+5)
LevelIterSeqSeekPrefixGE/skip=2/use-next=false-24                              687ns ± 2%     669ns ± 1%   -2.65%  (p=0.016 n=5+5)
LevelIterSeqSeekPrefixGE/skip=2/use-next=true-24                               167ns ± 1%     167ns ± 1%     ~     (p=0.730 n=5+5)
pkg:github.com/cockroachdb/pebble/sstable goos:linux goarch:amd64
BlockIterSeekGE/restart=16-24                                                  456ns ± 1%     456ns ± 0%     ~     (p=0.651 n=5+5)
BlockIterSeekLT/restart=16-24                                                  515ns ± 2%     519ns ± 5%     ~     (p=0.548 n=5+5)
BlockIterNext/restart=16-24                                                   19.2ns ± 0%    17.6ns ± 3%   -8.19%  (p=0.008 n=5+5)
BlockIterPrev/restart=16-24                                                   36.2ns ± 3%    36.7ns ± 1%     ~     (p=0.841 n=5+5)
TableIterSeekGE/restart=16,compression=Snappy-24                              1.46µs ± 1%    1.47µs ± 5%     ~     (p=0.794 n=5+5)
TableIterSeekGE/restart=16,compression=ZSTD-24                                1.51µs ± 6%    1.46µs ± 3%     ~     (p=0.151 n=5+5)
TableIterSeekLT/restart=16,compression=Snappy-24                              1.54µs ± 4%    1.54µs ± 5%     ~     (p=1.000 n=5+5)
TableIterSeekLT/restart=16,compression=ZSTD-24                                1.53µs ± 3%    1.54µs ± 3%     ~     (p=0.794 n=5+5)
TableIterNext/restart=16,compression=Snappy-24                                22.6ns ± 1%    22.3ns ± 2%     ~     (p=0.111 n=5+5)
TableIterNext/restart=16,compression=ZSTD-24                                  22.6ns ± 3%    22.4ns ± 2%     ~     (p=0.421 n=5+5)
TableIterPrev/restart=16,compression=Snappy-24                                42.9ns ± 1%    42.3ns ± 1%   -1.33%  (p=0.008 n=5+5)
TableIterPrev/restart=16,compression=ZSTD-24                                  43.1ns ± 1%    42.4ns ± 0%   -1.57%  (p=0.008 n=5+5)

name                                                                        old alloc/op   new alloc/op   delta
pkg:github.com/cockroachdb/pebble goos:linux goarch:amd64
IteratorSeekGE-24                                                              0.00B          0.00B          ~     (all equal)
IteratorNext-24                                                                0.00B          0.00B          ~     (all equal)
IteratorPrev-24                                                                0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=1/bloom=false/with-tombstone=false-24     0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=1/bloom=false/with-tombstone=true-24      0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=1/bloom=true/with-tombstone=false-24      0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=1/bloom=true/with-tombstone=true-24        495B ± 0%      495B ± 0%     ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=2/bloom=false/with-tombstone=false-24     0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=2/bloom=false/with-tombstone=true-24      0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=2/bloom=true/with-tombstone=false-24      0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=2/bloom=true/with-tombstone=true-24        495B ± 0%      495B ± 0%     ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=4/bloom=false/with-tombstone=false-24     0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=4/bloom=false/with-tombstone=true-24      0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=4/bloom=true/with-tombstone=false-24      0.00B          0.00B          ~     (all equal)
IteratorSeqSeekPrefixGENotFound/skip=4/bloom=true/with-tombstone=true-24        495B ± 0%      495B ± 0%     ~     (all equal)
IteratorSeqSeekGEWithBounds-24                                                 0.00B          0.00B          ~     (all equal)
IteratorSeekGENoop/withLimit=false-24                                          0.00B          0.00B          ~     (all equal)
IteratorSeekGENoop/withLimit=true-24                                           0.00B          0.00B          ~     (all equal)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward-24                     621B ± 8%      655B ±10%     ~     (p=0.127 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/backward-24                  1.70kB ± 4%    1.72kB ± 4%     ~     (p=0.222 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward-24                     643B ± 1%      646B ± 8%     ~     (p=0.222 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/backward-24                  1.72kB ± 0%    1.70kB ± 3%     ~     (p=1.000 n=4+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward-24                     558B ± 4%      572B ± 4%     ~     (p=0.222 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/backward-24                  1.61kB ± 2%    1.65kB ± 0%   +2.30%  (p=0.008 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward-24                    504B ± 1%      519B ± 1%   +3.02%  (p=0.008 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/backward-24                 1.57kB ± 0%    1.58kB ± 0%   +1.05%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=1,key-types=points-only-24                         64.0B ± 0%     80.0B ± 0%  +25.00%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=1,key-types=points-and-ranges-24                   64.0B ± 0%     80.0B ± 0%  +25.00%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-only-24                          192B ± 0%      224B ± 0%  +16.87%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-and-ranges-24                    192B ± 0%      224B ± 0%  +16.67%  (p=0.016 n=5+4)
IteratorScan/keys=100,r-amp=7,key-types=points-only-24                          449B ± 0%      512B ± 0%  +14.22%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=7,key-types=points-and-ranges-24                    449B ± 0%      513B ± 0%  +14.25%  (p=0.016 n=5+4)
IteratorScan/keys=100,r-amp=10,key-types=points-only-24                         642B ± 0%      771B ± 0%  +20.09%  (p=0.016 n=5+4)
IteratorScan/keys=100,r-amp=10,key-types=points-and-ranges-24                   642B ± 0%      771B ± 0%  +20.03%  (p=0.016 n=4+5)
IteratorScan/keys=1000,r-amp=1,key-types=points-only-24                        65.0B ± 0%     81.0B ± 0%  +24.62%  (p=0.016 n=4+5)
IteratorScan/keys=1000,r-amp=1,key-types=points-and-ranges-24                  65.0B ± 0%     80.6B ± 1%  +24.00%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-only-24                         194B ± 1%      227B ± 0%  +16.92%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-and-ranges-24                   194B ± 0%      225B ± 1%  +16.08%  (p=0.016 n=4+5)
IteratorScan/keys=1000,r-amp=7,key-types=points-only-24                         450B ± 0%      513B ± 0%  +14.14%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=7,key-types=points-and-ranges-24                   450B ± 0%      514B ± 0%  +14.17%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=10,key-types=points-only-24                        644B ± 0%      772B ± 0%  +19.84%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=10,key-types=points-and-ranges-24                  643B ± 0%      771B ± 0%  +19.94%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=1,key-types=points-only-24                       73.0B ± 8%     91.0B ± 0%  +24.66%  (p=0.016 n=5+4)
IteratorScan/keys=10000,r-amp=1,key-types=points-and-ranges-24                 75.0B ± 0%     88.4B ± 5%  +17.87%  (p=0.016 n=4+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-only-24                        210B ± 0%      244B ± 0%  +15.97%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-and-ranges-24                  206B ± 3%      244B ± 0%  +18.57%  (p=0.016 n=5+4)
IteratorScan/keys=10000,r-amp=7,key-types=points-only-24                        472B ± 0%      537B ± 0%  +13.82%  (p=0.029 n=4+4)
IteratorScan/keys=10000,r-amp=7,key-types=points-and-ranges-24                  463B ± 2%      534B ± 2%  +15.47%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-only-24                       667B ± 0%      796B ± 0%  +19.39%  (p=0.016 n=4+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-and-ranges-24                 667B ± 0%      789B ± 1%  +18.26%  (p=0.016 n=4+5)
CombinedIteratorSeek/range-key=false/batch=false-24                            64.0B ± 0%     80.0B ± 0%  +25.00%  (p=0.008 n=5+5)
CombinedIteratorSeek/range-key=false/batch=true-24                              128B ± 0%      144B ± 0%  +12.50%  (p=0.008 n=5+5)
CombinedIteratorSeek/range-key=true/batch=false-24                              377B ± 0%      393B ± 0%   +4.24%  (p=0.016 n=5+4)
CombinedIteratorSeek/range-key=true/batch=true-24                               441B ± 0%      457B ± 0%   +3.72%  (p=0.016 n=4+5)
CombinedIteratorSeek_Bounded-24                                                 368B ± 0%      384B ± 0%   +4.40%  (p=0.008 n=5+5)
CombinedIteratorSeekPrefix-24                                                   585B ± 0%      602B ± 0%   +2.91%  (p=0.008 n=5+5)
LevelIterSeekGE/restart=16/count=5-24                                          0.00B          0.00B          ~     (all equal)
LevelIterSeqSeekGEWithBounds/restart=16/count=5-24                             0.00B          0.00B          ~     (all equal)
LevelIterSeqSeekPrefixGE/skip=1/use-next=false-24                              0.00B          0.00B          ~     (all equal)
LevelIterSeqSeekPrefixGE/skip=1/use-next=true-24                               0.00B          0.00B          ~     (all equal)
LevelIterSeqSeekPrefixGE/skip=2/use-next=false-24                              0.00B          0.00B          ~     (all equal)
LevelIterSeqSeekPrefixGE/skip=2/use-next=true-24                               0.00B          0.00B          ~     (all equal)
pkg:github.com/cockroachdb/pebble/sstable goos:linux goarch:amd64
BlockIterSeekGE/restart=16-24                                                  0.00B          0.00B          ~     (all equal)
BlockIterSeekLT/restart=16-24                                                  0.00B          0.00B          ~     (all equal)
BlockIterNext/restart=16-24                                                    0.00B          0.00B          ~     (all equal)
BlockIterPrev/restart=16-24                                                    0.00B          0.00B          ~     (all equal)
TableIterSeekGE/restart=16,compression=Snappy-24                               0.00B          0.00B          ~     (all equal)
TableIterSeekGE/restart=16,compression=ZSTD-24                                 0.00B          0.00B          ~     (all equal)
TableIterSeekLT/restart=16,compression=Snappy-24                               0.00B          0.00B          ~     (all equal)
TableIterSeekLT/restart=16,compression=ZSTD-24                                 0.00B          0.00B          ~     (all equal)
TableIterNext/restart=16,compression=Snappy-24                                 0.00B          0.00B          ~     (all equal)
TableIterNext/restart=16,compression=ZSTD-24                                   0.00B          0.00B          ~     (all equal)
…